Ping Rewrite
Ping every client every time, not only the selectedClient, and not only
when the selectedClient disconnects.

Disconnect the client if it has closed the output(where we send
commands).

Iterate over again and select a client if selectedClient disconnected
and a client wasn't be selected in the current iteration.

Iterate over again and select another client if a
ConcurrentMidificationException is thrown.

Merges #3
Signed-off-by: Jadon Fowler <jadonflower@gmail.com>